modify metode addSpriteAnimation

+++ src/main/java/escapefromuniversity/inGame/GameView.java
@@ -40,7 +40,7 @@ public interface GameView {
      * @param height
      * @param width 
      */
-    void addSpriteAnimation(int id, State state, GameObjectType type, int height, int width, Point2D position);
+    void addSpriteAnimation(int id, State state, GameObjectType type, double height, double width, Point2D position);
 
     /**
      *
